24.12.17 enhance the service layer by adding the validation logic to ensure data integrity

+++ spring-project-hixview/src/main/java/site/hixview/jpa/service/PressEntityService.java
@@ -2,13 +2,20 @@ package site.hixview.jpa.service;
 
 import jakarta.persistence.EntityExistsException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.hixview.aggregate.domain.Press;
 import site.hixview.aggregate.error.EntityNotFoundWithNumberException;
 import site.hixview.aggregate.service.PressService;
+import site.hixview.jpa.entity.CompanyArticleEntity;
+import site.hixview.jpa.entity.EconomyArticleEntity;
+import site.hixview.jpa.entity.IndustryArticleEntity;
 import site.hixview.jpa.entity.PressEntity;
 import site.hixview.jpa.mapper.PressEntityMapperImpl;
+import site.hixview.jpa.repository.CompanyArticleEntityRepository;
+import site.hixview.jpa.repository.EconomyArticleEntityRepository;
+import site.hixview.jpa.repository.IndustryArticleEntityRepository;
 import site.hixview.jpa.repository.PressEntityRepository;
 
 import java.util.List;
@@ -16,6 +23,7 @@ import java.util.Optional;
 
 import static site.hixview.aggregate.util.ExceptionUtils.getFormattedExceptionMessage;
 import static site.hixview.aggregate.vo.ExceptionMessage.ALREADY_EXISTED_ENTITY;
+import static site.hixview.aggregate.vo.ExceptionMessage.REMOVE_REFERENCED_ENTITY;
 import static site.hixview.aggregate.vo.WordCamel.ENGLISH_NAME;
 import static site.hixview.aggregate.vo.WordCamel.NUMBER;
 
@@ -25,6 +33,9 @@ import static site.hixview.aggregate.vo.WordCamel.NUMBER;
 public class PressEntityService implements PressService {
 
     private final PressEntityRepository pressEntityRepository;
+    private final CompanyArticleEntityRepository companyArticleEntityRepository;
+    private final IndustryArticleEntityRepository industryArticleEntityRepository;
+    private final EconomyArticleEntityRepository economyArticleEntityRepository;
     private final PressEntityMapperImpl mapper = new PressEntityMapperImpl();
 
     @Override
@@ -75,7 +86,23 @@ public class PressEntityService implements PressService {
             throw new EntityExistsException(getFormattedExceptionMessage(
                     ALREADY_EXISTED_ENTITY, ENGLISH_NAME, englishName, Press.class));
         }
-        return mapper.toPress(pressEntityRepository.save(mapper.toPressEntity(press)));
+        PressEntity pressEntity = pressEntityRepository.save(mapper.toPressEntity(press));
+        companyArticleEntityRepository.saveAll(
+                companyArticleEntityRepository.findByPress(pressEntity).stream().map(article -> {
+            article = CompanyArticleEntity.builder().companyArticle(article).press(pressEntity).build();
+            return article;
+        }).toList());
+        economyArticleEntityRepository.saveAll(
+                economyArticleEntityRepository.findByPress(pressEntity).stream().map(article -> {
+            article = EconomyArticleEntity.builder().economyArticle(article).press(pressEntity).build();
+            return article;
+        }).toList());
+        industryArticleEntityRepository.saveAll(
+                industryArticleEntityRepository.findByPress(pressEntity).stream().map(article -> {
+            article = IndustryArticleEntity.builder().industryArticle(article).press(pressEntity).build();
+            return article;
+        }).toList());
+        return mapper.toPress(pressEntity);
     }
 
     @Override
@@ -83,6 +110,13 @@ public class PressEntityService implements PressService {
     public void removeByNumber(Long number) {
         if (!pressEntityRepository.existsByNumber(number)) {
             throw new EntityNotFoundWithNumberException(number, Press.class);
+        }
+        PressEntity pressEntity = pressEntityRepository.findByNumber(number).orElseThrow();
+        if (!companyArticleEntityRepository.findByPress(pressEntity).isEmpty() ||
+                !economyArticleEntityRepository.findByPress(pressEntity).isEmpty() ||
+                !industryArticleEntityRepository.findByPress(pressEntity).isEmpty()) {
+            throw new DataIntegrityViolationException(getFormattedExceptionMessage(
+                    REMOVE_REFERENCED_ENTITY, NUMBER, number, PressEntity.class));
         }
         pressEntityRepository.deleteByNumber(number);
     }
